+++ lib/vissen/output/context/circle.rb
@@ -8,11 +8,11 @@ module Vissen
       # points of the element array.
       class Circle < Cloud
         # @param  point_count [Integer] the number of points.
-        # @param  offset [Float] the angle offset, in radians, of the first
+        # @param  offset [Numeric] the angle offset, in radians, of the first
         #   point.
-        # @param  width [Float] (see Context)
-        # @param  height [float] (see Context)
-        # @param  radius [Float] the radius of the context.
+        # @param  width [Numeric] (see Context)
+        # @param  height [Numeric] (see Context)
+        # @param  radius [Numeric] the radius of the context.
         # @param  args (see CloudContext).
         def initialize(point_count,
                        offset: 0,