+++ lib/vissen/output/filter.rb
@@ -23,7 +23,7 @@ module Vissen
       # This method should apply the filter to the given `PixelBuffer`.
       #
       # @raise  NotImplementedError if the method is not implemented in the
-      # specific `Filter` implementation.
+      #   specific `Filter` implementation.
       #
       # @param  _pixel_buffer [PixelBuffer] the pixel cloud to which the filter
       #   should be applied.