+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.6.1] - 2018-04-20
+### Changed
+- Improved, more descriptive comments.
+
 ## [0.6.0] - 2018-04-14
 ### Added
 - Output filter support in PixelBuffer.